Moved service locator to app package
